+++ resources/views/bootstrap-5/includes/bulk-actions.blade.php
@@ -1,7 +1,7 @@
 @if (count($bulkActions))
     <div class="dropdown mb-3 mb-md-0 d-block d-md-inline">
         <button class="btn dropdown-toggle d-block w-100 d-md-inline" type="button" id="bulkActions" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-            {{ __('Bulk Actions') }}
+            @lang('Bulk Actions')
         </button>
 
         <div class="dropdown-menu dropdown-menu-end w-100" aria-labelledby="bulkActions">