+++ resources/views/bootstrap-5/includes/table.blade.php
@@ -31,7 +31,12 @@
                 <x-livewire-tables::bs5.table.cell colspan="{{ count($bulkActions) ? count($columns) + 1 : count($columns) }}">
                     @unless ($selectAll)
                         <div>
-                            <span>{!! __('You have selected <strong>:count</strong> rows, do you want to select all <strong>:total</strong>?', ['count' => $rows->count(), 'total' => number_format($rows->total())]) !!}</span>
+                            <span>
+                                @lang('You have selected')
+                                <strong>{{ $rows->count() }}</strong>
+                                @lang('rows, do you want to select all')
+                                <strong>{{ number_format($rows->total()) }}</strong>?
+                            </span>
 
                             <button
                                 wire:click="selectAll"
@@ -43,7 +48,11 @@
                         </div>
                     @else
                         <div>
-                            <span>{!! __('You are currently selecting all <strong>:total</strong> rows.', ['total' => number_format($rows->total())]) !!}</span>
+                            <span>
+                                @lang('You are currently selecting all')
+                                <strong>{{ number_format($rows->total()) }}</strong>
+                                @lang('rows').
+                            </span>
 
                             <button
                                 wire:click="resetBulk"
@@ -81,7 +90,7 @@
         @empty
             <x-livewire-tables::bs5.table.row>
                 <x-livewire-tables::bs5.table.cell colspan="{{ count($bulkActions) ? count($columns) + 1 : count($columns) }}">
-                    @lang('No items found. Try narrowing your search.')
+                    @lang($emptyMessage)
                 </x-livewire-tables::bs5.table.cell>
             </x-livewire-tables::bs5.table.row>
         @endforelse