+++ resources/views/tailwind/includes/table.blade.php
@@ -81,7 +81,7 @@
                         <div class="flex rounded-md shadow-sm">
                             <input
                                 wire:model="selected"
-                                value="{{ $row->getKey() }}"
+                                value="{{ $row->{$primaryKey} }}"
                                 onclick="event.stopPropagation();return true;"
                                 type="checkbox"
                                 class="rounded-md shadow-sm border-gray-300 block transition duration-150 ease-in-out sm:text-sm sm:leading-5"