+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to `laravel-livewire-tables` will be documented in this file
 
 - Modified views to support localization better where necessary (republish views).
 - Alphabetize en.json
+- Fixed bulk actions using wrong key to select instead of $primaryKey
 
 ## [1.3.1] - 2021-04-26