+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to `laravel-livewire-tables` will be documented in this file
 - Modified views to support localization better where necessary (republish views).
 - Alphabetize en.json
 - Fixed bulk actions using wrong key to select instead of $primaryKey
+- Make bulk select checkbox use primary key
 
 ## [1.3.1] - 2021-04-26